Add CLI help snapshot tests.
