+++ libraries/ts-command-line/src/test/CommandLineParameter.test.ts
@@ -2,13 +2,13 @@
 // See LICENSE in the project root for license information.
 
 import * as argparse from 'argparse';
-import { AnsiEscape } from '@rushstack/terminal';
 
 import { DynamicCommandLineParser } from '../providers/DynamicCommandLineParser';
 import { DynamicCommandLineAction } from '../providers/DynamicCommandLineAction';
 import { CommandLineParameterBase } from '../parameters/BaseClasses';
 import type { CommandLineParser } from '../providers/CommandLineParser';
 import type { CommandLineAction } from '../providers/CommandLineAction';
+import { ensureHelpTextMatchesSnapshot } from './helpTestUtilities';
 
 interface IExtendedArgumentParser extends argparse.ArgumentParser {
   _printMessage: (message: string) => void;
@@ -174,18 +174,9 @@ describe(CommandLineParameterBase.name, () => {
     process.env = existingEnv;
   });
 
-  it('prints the global help', () => {
+  it('renders help text', () => {
     const commandLineParser: CommandLineParser = createParser();
-    const helpText: string = AnsiEscape.removeCodes(commandLineParser.renderHelpText());
-    expect(helpText).toMatchSnapshot();
-  });
-
-  it('prints the action help', () => {
-    const commandLineParser: CommandLineParser = createParser();
-    const helpText: string = AnsiEscape.removeCodes(
-      commandLineParser.getAction('do:the-job').renderHelpText()
-    );
-    expect(helpText).toMatchSnapshot();
+    ensureHelpTextMatchesSnapshot(commandLineParser);
   });
 
   it('parses an input with ALL parameters', async () => {
