+++ libraries/ts-command-line/src/test/DynamicCommandLineParser.test.ts
@@ -4,6 +4,7 @@
 import { DynamicCommandLineParser } from '../providers/DynamicCommandLineParser';
 import { DynamicCommandLineAction } from '../providers/DynamicCommandLineAction';
 import type { CommandLineFlagParameter } from '../parameters/CommandLineFlagParameter';
+import { ensureHelpTextMatchesSnapshot } from './helpTestUtilities';
 
 describe(DynamicCommandLineParser.name, () => {
   it('parses an action', async () => {
@@ -22,6 +23,8 @@ describe(DynamicCommandLineParser.name, () => {
       parameterLongName: '--flag',
       description: 'The flag'
     });
+
+    ensureHelpTextMatchesSnapshot(commandLineParser);
 
     await commandLineParser.executeAsync(['do:the-job', '--flag']);
